Worked on sections: 1. Creating an app 2. Components 3. Collections
I like this framework so far, though some of it resources are a bit old
and I'm wondering whether Next.js would be a better future-proof option.

On the other hand, I like how it comes with some cool features out of
the box, the collections stuff reactivity and Mongo integration is
pretty dope.